requirements.txt
opencv-python>=4.5.0
facenet-pytorch>=2.5.0       # Replaces mtcnn
mediapipe>=0.8.0
torch>=1.7.0
torchvision>=0.8.0
flask>=2.0.0
pyyaml>=5.0.0
numpy>=1.20.0